Created class level for loading tmx maps from Tiled Map Editor. Fixed button's line thickness, when changing scale. Blurry text also fixed. Ui brokes, when resizing window. Will be fixed after finishing object loading in class Level. Created some tiled map editor's levels for testing class Level. Created class locale for manipulating with languages. Will be written in next releases.

+++ code/ui/button.hpp
@@ -29,7 +29,7 @@ namespace UINAMESPACE
         void draw(sf::RenderTarget& target, sf::RenderStates states) const;
 
     private:
-        float b_text_char_size;
+        unsigned int b_text_char_size;
         sf::RectangleShape b_fone;
         LerpingColor       b_fone_color;
         LerpingColor       b_outline_color;
